User/ivberg/perfetto process (#91)
* Bring GUI Table Process support with arg fields & process label from Chromium. 
* Enhance GenericEvents to show this process label info if available.
* Add Android Package table
* Add rowcount logging for Perfetto sql queries. 
* Slightly better Unit Tests for new tables by building tables as well (data not processed yet)

+++ PerfettoCds/Pipeline/DataOutput/PerfettoFtraceEvent.cs.cs
@@ -15,7 +15,7 @@ namespace PerfettoCds.Pipeline.DataOutput
         public string ProcessFormattedName { get; }
         public string ThreadFormattedName { get; }
         public string ThreadName { get; }
-        public long Tid { get; }
+        public uint Tid { get; }
         public uint Cpu { get; }
         // Name of the ftrace event
         public string Name { get; }
@@ -28,7 +28,7 @@ namespace PerfettoCds.Pipeline.DataOutput
             string processFormattedName,
             string threadFormattedName,
             string threadName,
-            long tid,
+            uint tid,
             uint cpu,
             string name,
             List<string> values,
